refactor(capability)!: rename contract module
BREAKING CHANGE: github.com/go-fries/fries/contract/v4 has moved to github.com/go-fries/fries/capability/v4.
