+++ capability/countable.go
@@ -1,4 +1,4 @@
-package contract
+package capability
 
 type Countable interface {
 	Count() int
