+++ capability/go.mod
@@ -1,0 +1,3 @@
+module github.com/go-fries/fries/capability/v4
+
+go 1.25.0
