+++ capability/htmlable.go
@@ -1,4 +1,4 @@
-package contract
+package capability
 
 type Htmlable interface {
 	Html() string
