+++ capability/jsonable.go
@@ -1,4 +1,4 @@
-package contract
+package capability
 
 type Jsonable interface {
 	Json() string
