+++ capability/nameable.go
@@ -1,4 +1,4 @@
-package contract
+package capability
 
 type Nameable interface {
 	Name() string
