+++ capability/renderable.go
@@ -1,4 +1,4 @@
-package contract
+package capability
 
 type Renderable interface {
 	Render() string
