+++ capability/retriable.go
@@ -1,4 +1,4 @@
-package contract
+package capability
 
 type Retriable interface {
 	Retries() int
