+++ capability/server.go
@@ -1,4 +1,4 @@
-package contract
+package capability
 
 import "context"
