--- contract/go.mod
@@ -1,3 +1,0 @@
-module github.com/go-fries/fries/contract/v4
-
-go 1.25.0
