+++ http/server/ignore_closed.go
@@ -5,20 +5,20 @@ import (
 	"errors"
 	"net/http"
 
-	"github.com/go-fries/fries/contract/v4"
+	"github.com/go-fries/fries/capability/v4"
 )
 
 // IgnoreServerClosed returns a server that treats [http.ErrServerClosed] from Start as nil.
-func IgnoreServerClosed(srv contract.Server) *IgnoreClosedServer {
+func IgnoreServerClosed(srv capability.Server) *IgnoreClosedServer {
 	return &IgnoreClosedServer{srv: srv}
 }
 
 // IgnoreClosedServer wraps a server and ignores [http.ErrServerClosed] returned by Start.
 type IgnoreClosedServer struct {
-	srv contract.Server
+	srv capability.Server
 }
 
-var _ contract.Server = (*IgnoreClosedServer)(nil)
+var _ capability.Server = (*IgnoreClosedServer)(nil)
 
 // Start starts the wrapped server and returns nil for [http.ErrServerClosed].
 func (s *IgnoreClosedServer) Start(ctx context.Context) error {
